requirements.txt
asgiref==3.2.7
astroid==2.4.0
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-braces==1.14.0
django-phonenumber-field==4.0.0
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
phonenumbers==8.12.3
psycopg2-binary==2.8.5
PyJWT==1.7.1
pylint==2.5.0
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
twilio==6.38.1
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
